rgw/sfs: Set the type of the default user to TYPE_RGW.
Right now the type of our default user is TYPE_NONE which seems to be incorrect because new users that are created with the AdminOps API have the type TYPE_RGW. The same is when creating users via the CLI, e.g. 'radosgw-admin user create --uid=dev --display-name=Developer --system' (in a SES7.1 cluster).

Signed-off-by: Volker Theile <vtheile@suse.com>