rgw/sfs: Initial Garbage collector thread.
This adds an initial GC approach just starting a thread and running
 a basic remove deleted buckets operation.

The implementation of how the thread is started and stopped are taken
from rgw's code for consistency across stores.

It uses 2 parameters from rgw:

- `rgw_gc_processor_period` : defines the time GC waits to be triggered
- `rgw_gc_max_objs` : defines the maximum number of items the GC can
  remove per iteration.

Any possible actions done to the `sqlite` metadata is considered an item
(object) when taking into account the maxium number of items to be
processed in a single iteration as those actions block the metadata
database.

Fixes: https://github.com/aquarist-labs/s3gw/issues/165
Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>