Add s3gw specfile, systemd service and sysconfig file
This commit adds an s3gw.spec file for building a standalone
s3gw RPM, along with a systemd unit file and sysconfig file to
tweak settings (e.g. local data directory).  The systemd unit
file and sysconfig file are not necessary or even desirable
when running containerized, but make sense in the context of
a standalone package.

s3gw.spec is based on the existing ceph.spec.in, but instead of
creating a ceph package plus a bunch of library packages and
other subpackages, s3gw.spec creates a single RPM containing the
radosgw binary, librados, libradosgw and libceph-common.  The
s3gw RPM is set to conflict with ceph-radosgw and librados2,
to avoid accidental installation alongside (or on top of!)
a regular ceph deployment.

With this, we can do package builds on OBS with pretty much
nothing other than a _service file, as can be currently seen at
https://build.opensuse.org/package/show/home:tserong:branches:filesystems:ceph:s3gw/s3gw

<?xml version="1.0"?>
<services>
  <service name="obs_scm">
    <param name="versionformat">0.7.0.@PARENT_TAG@.@TAG_OFFSET@</param>
    <param name="url">https://github.com/tserong/ceph.git</param>
    <param name="scm">git</param>
    <param name="revision">s3gw-obs-automation-experiment</param>
    <param name="extract">s3gw.spec</param>
    <param name="changesgenerate">enable</param>
  </service>
  <service name="set_version"/>
  <service name="tar" mode="buildtime"/>
</services>

Given the above, OBS will go grab the source from github, extract
s3gw.spec, plug in a somewhat sensible version number, generate a
somewhat useful changelog and build an RPM for us.

Note:

- The primary purpose of this work is to enable automatic builds
  on build.opensuse.org.  That's why the version in the spec file
  is set to "WILL_BE_OVERWRITTEN_BY_OBS".
- There are some SUSE-isms in the spec file, especially in the
  %pre and %post scripts.  This will need adjusting if we want
  to try Fedora/RH builds.
- s3gw is configured to use system boost, with necessary build
  dependencies available from filesystems:ceph:s3gw on OBS.
- I've switched to using ninja instead of make for the build,
  because I couldn't get make to only build radsogw and deps.
- I've set -DENABLE_GIT_VERSION=OFF, because this allows us to
  do builds straight from a git checkout, without needing to
  generate a source tarball with `make-dist`.
- There may be other cmake flags that we can still turn off, or
  other things that can be stripped back further.

Fixes: https://github.com/aquarist-labs/s3gw/issues/154
Signed-off-by: Tim Serong <tserong@suse.com>