ci/fix: Keep the `.note.ABI-tag` section
Keep the `.note.*` section when stripping the binaries during a release
build.

The LSB specifies that a binary shall contain a `.note.ABI-tag` section,
containing information about the expected kernel/system ABI [1]. On most
systems this is not enforced, but some (like Oracle 9) enforce this and
missing this section causes the binary to fail execution with and "exec
format error".

[1]: https://refspecs.linuxfoundation.org/LSB_1.2.0/gLSB/noteabitag.html

Fixes: aquarist-labs/s3gw#305

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>