ceph.spec.in: Require Cython >= 0.29 but < 3
This is based on a change made to fix the build on openSUSE
Factory which now includes Cython 0.29.x and Cython 3:

  https://build.opensuse.org/request/show/1104240

Ceph currently only builds with Cython 0.29, hence the need
to exclude version 3 which would otherwise be used by default.

Unfortunately the fix for Factory doesn't work when building
against SLE 15 SP3, due to older python macros, hence this
version.

Signed-off-by: Tim Serong <tserong@suse.com>