ceph.spec.in: Require fmt-devel < 10 (bsc#1213217)
openSUSE Factory is upgrading to fmt version 10, but Ceph
fails to build with that.  A separate package (fmt-9) has
been created in Factory so we can use that instead.  The
trick here is to set the BuildRequires to it will use the
regular fmt-devel package if a version < 10 is available
(e.g. on SLE 15 SP3 for SES 7.1), but to use fmt-9-devel
otherwise.

Signed-off-by: Tim Serong <tserong@suse.com>