rgw/sfs: Change types.h Object ownership story
Object was a shared_ptr hidden behind 'ObjectRef' typedef. The shared
ptr stems back from when there was a map of objects cache. Since we
are no longer sharing Objects remove the shared_ptr and make the
ownership of Object explicit. In most cases by using a unique_ptr.

This also changes the behavior of refresh_meta when there is not (yet)
a database record. We used to keep objref at nullptr. Now we
initialize it with a default object that has deleted = true and set
our state.exists to false.

Signed-off-by: Marcel Lauhoff <marcel.lauhoff@suse.com>